feature(CustomSpecificationBuilder.java) add custom spec builder
