--- criteria-builder/src/main/resources/application.yml
@@ -1,51 +1,0 @@
-server:
-  port: 8082
-
-logging:
-  file:
-    max-size: 100MB
-    path: .
-
-  pattern:
-    console: "%clr(%date{HH:mm:ss.SSS}) %level %clr(%logger{0}) : %message%n"
-    file:    "%clr(%date) %level %clr(%logger :) %message%n"
-
-  level:
-    root: INFO
-    org.springframework.web: error
-
-spring:
-  output:
-    ansi:
-      enabled: ALWAYS
-
-  main:
-    allow-bean-definition-overriding: true
-
-  datasource:
-    platform: postgres
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:testdb
-    username: user
-    password: password
-
-  jpa:
-    database-platform=org.hibernate.dialect.H2Dialect:
-
-    properties:
-      hibernate:
-        temp:
-          use_jdbc_metadata_defaults: false
-        dialect: org.hibernate.dialect.PostgreSQL95Dialect
-        current_session_context_class: org.springframework.orm.hibernate5.SpringSessionContext
-      show-sql: true
-
-    hibernate:
-      ddl-auto: create-drop
-    open-in-view: false
-    show-sql: true
-
-
-  h2:
-    console:
-      enabled: true
