+++ query-dsl/pom.xml
@@ -11,7 +11,7 @@
         <relativePath>../</relativePath>
     </parent>
 
-    <artifactId>org.restful.querying.instances.criteria.builder.query.dsl</artifactId>
+    <artifactId>query-dsl</artifactId>
     <name>query-dsl</name>
     <description>
         This module presents the simple instances of the restful service query method using QueryDSL.
@@ -19,6 +19,7 @@
 
     <properties>
         <querydsl.version>4.2.2</querydsl.version>
+        <start-class>org.restful.querying.instances.query.dsl.QueryDsl</start-class>
     </properties>
 
     <dependencies>
