--- query-dsl/src/main/java/org/restful/querying/instances/MainQueryDsl.java
@@ -1,21 +1,0 @@
-package org.restful.querying.instances;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-import org.springframework.data.web.config.EnableSpringDataWebSupport;
-
-@EnableSpringDataWebSupport
-@SpringBootApplication
-public class MainQueryDsl extends SpringBootServletInitializer {
-
-    public static void main(String[] args) throws Exception {
-        SpringApplication.run(MainQueryDsl.class, args);
-    }
-
-    @Override
-    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-        return application.sources(MainQueryDsl.class);
-    }
-}
