+++ query-dsl/src/main/java/org/restful/querying/instances/query/dsl/QueryDsl.java
@@ -1,0 +1,21 @@
+package org.restful.querying.instances.specification.builder;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+import org.springframework.data.web.config.EnableSpringDataWebSupport;
+
+@EnableSpringDataWebSupport
+@SpringBootApplication
+public class QueryDsl extends SpringBootServletInitializer {
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(QueryDsl.class, args);
+    }
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(QueryDsl.class);
+    }
+}
