+++ specification-builder/pom.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <artifactId>restful-querying-instances</artifactId>
+        <groupId>org.restful.querying.instances</groupId>
+        <version>${revision}${sha1}${changelist}</version>
+        <relativePath>../</relativePath>
+    </parent>
+
+    <artifactId>specification-builder</artifactId>
+    <name>specification-builder</name>
+    <description>
+        This module presents the simple instances of the restful service query method using Dynamic Criteria Builder.
+    </description>
+
+    <properties>
+        <start-class>org.restful.querying.instances.specification.builder.SpecificationBuilder</start-class>
+    </properties>
+
+</project>
