+++ specification-builder/src/main/java/org/restful/querying/instances/specification/builder/SpecificationBuilder.java
@@ -1,0 +1,19 @@
+package org.restful.querying.instances.specification.builder;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+@SpringBootApplication
+public class SpecificationBuilder extends SpringBootServletInitializer {
+
+    public static void main(String[] args) throws Exception {
+        SpringApplication.run(SpecificationBuilder.class, args);
+    }
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(SpecificationBuilder.class);
+    }
+}
